pyproject.toml
[tool.poetry]
name = "mlops-project"
version = "0.1.0"
description = "MLops project OmegaLUL"
authors = ["M1croZavr <tlashukov@gmail.com>"]
repository = "https://github.com/M1croZavr/mlops"
readme = "README.md"
package-mode = true
packages = [
    {include = "dashboard", from = "src"},
    {include = "fastapi_service", from = "src"},
    {include = "grpc_service", from = "src"},
    {include = "models", from = "src"}
]

[tool.poetry.dependencies]
python = "^3.10"
pytorch-lightning = "^2.4.0"
fastapi = {extras = ["standard"], version = "^0.115.4"}
torchvision = "^0.20.1"
python-multipart = "^0.0.17"
idx2numpy = "^1.2.3"
grpcio = "^1.67.1"
grpcio-tools = "^1.67.1"
streamlit = "^1.40.1"
streamlit-drawable-canvas = "^0.9.3"
psutil = "^6.1.0"

[tool.poetry.scripts]
start-server = "src.fastapi_service.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
